refactoring admin code into separate components in ./includes. Using Material-UI to update grid system and add Cards. Added navbar to admin. styled buttons
