+++ src/styles/AdminView.scss
@@ -1,21 +1,61 @@
 .AdminView {
+    position: relative;
     .course-container {
-        display: grid;
-        grid-template-columns: auto auto auto auto;
-        grid-gap: 5%;
+        width: 100%;
+        display: flex;
         flex-direction: row;
-        justify-content: center;
-        align-content: center;
+        align-items: stretch;
+        padding: 2em;
+    }
+
+    .create-course {
+        display: inline-block;
+        background-color: white;
+        border: solid 1px #484848;
+        margin: 5px;
+        width: 150px;
+        height: 40px;
+        border-radius: 4px;
+        text-decoration: none;
+        cursor: pointer;
     }
 
     .course {
         width: 100%;
         margin: 1em auto;
         padding: 1em;
-        border: 1px solid #ccc;
+        min-height: 100%;
 
         .course-section {
             margin: 1em;
         }
+
+        .editing-button {
+            display: inline-block;
+            background-color: white;
+            border: solid 1px #484848;
+            border-radius: 4px;
+            text-decoration: none;
+            padding: .5em;
+            margin: .25em;
+            cursor: pointer;
+        }
+    
+        .roles-button {
+            display: inline-block;
+            background-color: white;
+            border: solid 1px #484848;
+            border-radius: 4px;
+            text-decoration: none;
+            padding: .5em;
+            margin: .25em;
+            cursor: pointer;
+        }
+    }
+}
+
+@media(max-width: $mobile-breakpoint) {
+    .AdminView {
+        background-color: #f5f5f5;
     }
 }
